Wrapper & build updates cont'd

+++ src/main/bin/build_wrapper.sh
@@ -2,20 +2,25 @@
 
 VERSION=$1
 
-# On python 3
-path_to_pyenv=$(which pyenv)
- if [ -x "$path_to_executable" ] ; then
-    path_to_pyenv wrapper.build.env
-else
-    /usr/local/bin/virtualenv --distribute wrapper.build.venv
+if [[ -z $NO_VENV ]] ; then
+	# On python 3
+	path_to_pyenv=$(which pyenv)
+	 if [ -x "$path_to_executable" ] ; then
+	    path_to_pyenv wrapper.build.env
+	else
+	    /usr/local/bin/virtualenv --distribute wrapper.build.venv
+	fi
+
+	source wrapper.build.venv/bin/activate
+
+	pip install jcc
 fi
-
-source wrapper.build.venv/bin/activate
-
-pip install jcc
 
 python "../../src/main/python/build_wrapper.py" "$VERSION" ../../pom.xml "-Pjenkins"
 
-deactivate
 
-rm -rf wrapper.build.venv
+if [[ -z $NO_VENV ]] ; then
+	deactivate
+
+	rm -rf wrapper.build.venv
+fi
